RMMapView/locationManager: Remove old userLocations from annotations
There seems to have been a subtle change in the CoreLocation framework.  This
change resulted in errant 'legitimate' position reports.  If we don't clean
them up when we have new information, shit goes to hell real quick.

So, to fix this, we enforce the presumed rule[0] that there be only one
userLocation annotation in the annotations list at any given time.

[0] "presumed" because there's no documentation, and the change history leaves a
lot to be desired. :-/

Signed-off-by: Jason Cooper <jason@leadnavsystems.com>